+++ application/views/tool/toolnotice.php
@@ -1,4 +1,47 @@
+<div class="progressContainer">
+    <div class="progressBar"></div>
+</div>
+
 <main>
+    <!-- HEADING -->
+    <div class="d-none d-lg-block">
+        <div class="tool-heading">
+            <div class="tool-heading-text">
+                <div>
+                    <span class="tool-heading-type"><?php echo $tool_head['type']; ?></span>
+                </div>
+                <div>
+                    <span class="tool-heading-brand"><?php echo $tool_head['brand']; ?></span>
+                </div>
+            </div>
+            <div class="tool-heading-nav">
+                <div class="tool-heading-link">
+                    <a href="<?php echo base_url('tool/info/'); echo $tool_head['slug']; ?>">
+                        <span class="icon-tool-info"></span>
+                    </a>
+                </div>
+                <div class="tool-heading-link">
+                    <span class="icon-tool-notice active"></span>
+                </div>
+                <div class="tool-heading-link">   
+                    <a href="<?php echo base_url('tool/consumables/'); echo $tool_head['slug']; ?>">
+                        <span class="icon-tool-consumables"></span>
+                    </a>
+                </div>
+                <div class="tool-heading-link">
+                    <a href="<?php echo base_url('tool/help/'); echo $tool_head['slug']; ?>">
+                        <span class="icon-tool-help"></span>
+                    </a>
+                </div>
+                <div class="tool-heading-link">
+                    <a href="<?php echo base_url('tool/tips/'); echo $tool_head['slug']; ?>">
+                        <span class="icon-tool-tips"></span>
+                    </a>
+                </div>
+            </div>
+        </div>
+    </div>
+
     <!-- MOBILE MENU -->
     <div class="d-sm-block d-md-block d-lg-none d-xl-none">
         <div class="tool-menu-mobile">
@@ -37,62 +80,43 @@
         </div>
     </div>
 
-    <div class="container">
+    <div class="container mt-8">
         <div class="row">
             <!-- DESKTOP MENU -->
             <div class="d-none d-lg-block col-lg-4 col-xl-4">
-                <div class="tool-menu">
-                    <div class="tool-menu-head p-5">
-                        <div>
-                            <span class="tool-menu-type"><?php echo $tool_head['type']; ?></span>
-                        </div>
-                        <div>
-                            <span class="tool-menu-brand"><?php echo $tool_head['brand']; ?></span>
+                <div class="tool-notice-menu" id="toStick">
+                    <div class="notice-menu-part">
+                        <div class="notice-menu-item notice-menu-h1">
+                            <a href="#step1">Obtenir un modèle 3D</a>
                         </div>
                     </div>
-                    <div class="tool-menu-body p-5">
-                        <ul>
-                            <li class="tool-menu-link py-3">
-                                <a href="<?php echo base_url('tool/info/'); echo $tool_head['slug']; ?>" class="tool-link">
-                                    <div class="tool-link-item">
-                                        <div><span class="icons-info mr-4"></span></div>
-                                        <div><span class="tool-link-text">Informations</span></div>
-                                    </div>
-                                </a>
-                            </li>
-                            <li class="tool-menu-link py-3">
-                                <a href="#" class="tool-link active">
-                                    <div class="tool-link-item">
-                                        <div><span class="icons-notice mr-4"></span></div>
-                                        <div><span class="tool-link-text">Notice</span></div>
-                                    </div>
-                                </a>
-                            </li>
-                            <li class="tool-menu-link py-3">
-                                <a href="<?php echo base_url('tool/consumables/'); echo $tool_head['slug']; ?>" class="tool-link">
-                                    <div class="tool-link-item">
-                                        <div><span class="icons-consumable mr-4"></span></div>
-                                        <div><span class="tool-link-text">Consommables</span></div>
-                                    </div>
-                                </a>
-                            </li>
-                            <li class="tool-menu-link py-3">
-                                <a href="<?php echo base_url('tool/help/'); echo $tool_head['slug']; ?>" class="tool-link">
-                                    <div class="tool-link-item">
-                                        <div><span class="icons-faq mr-4"></span></div>
-                                        <div><span class="tool-link-text">FAQ</span></div>
-                                    </div>
-                                </a>
-                            </li>
-                            <li class="tool-menu-link py-3">
-                                <a href="<?php echo base_url('tool/tips/'); echo $tool_head['slug']; ?>" class="tool-link">
-                                    <div class="tool-link-item">
-                                        <div><span class="icons-tips mr-4"></span></div>
-                                        <div><span class="tool-link-text">Astuces</span></div>
-                                    </div>
-                                </a>
-                            </li>
-                        </ul>
+                    <div class="notice-menu-part">
+                        <div class="notice-menu-item notice-menu-h1">
+                            <a href="#step2">Préparer son impression</a>
+                        </div>
+                        <div class="notice-menu-item notice-menu-h2">
+                            <a href="#step2-1">Éditer votre objet</a>
+                        </div>
+                        <div class="notice-menu-item notice-menu-h2">
+                            <a href="#step2-2">Régler les paramètres d’impression</a>
+                        </div>
+                        <div class="notice-menu-item notice-menu-h2">
+                            <a href="#step2-3">Finaliser le fichier d’impression</a>
+                        </div>
+                    </div>
+                    <div class="notice-menu-part">
+                        <div class="notice-menu-item notice-menu-h1">
+                            <a href="#step3">Imprimer son objet</a>
+                        </div>
+                        <div class="notice-menu-item notice-menu-h2">
+                            <a href="#step3-1">Installer son filament</a>
+                        </div>
+                        <div class="notice-menu-item notice-menu-h2">
+                            <a href="#step3-2">Lancer son impression</a>
+                        </div>
+                        <div class="notice-menu-item notice-menu-h2">
+                            <a href="#step3-3">Récupération de l’impression</a>
+                        </div>
                     </div>
                 </div>
             </div>
@@ -122,7 +146,7 @@
                         <div class="p-6">
 
                             <div class="step-level-1">
-                                <h2 class="mb-4">Obtenir un modèle 3D</h2>
+                                <h2 id="step1" class="mb-4">Obtenir un modèle 3D</h2>
                                 <p>L’objet que vous souhaitez imprimer doit préalablement se présenter sous la forme d’un fichier contenant votre objet modélisé en 3D.</p>
                                 <p>Si votre objet est très spécifique, il vous faudra préalablement le dessiner en trois dimensions dans un logiciel de modélisation 3D. Gardez en tête que prendre en main le logiciel peut prendre un certain temps.</p>
                                 <p>Si votre objet est générique, vous pouvez peut-être en trouver un modèle déjà existant en ligne. À ce titre, il existe des plateformes qui permettent de partager et de télécharger des objets modélisés en 3D. Ces fichiers sont libres et proposés gratuitement par des contributeurs du monde entier. Parmi ces plateformes, citons <a href="https://www.thingiverse.com/">Thingiverse</a> ou <a href="https://cults3d.com/fr">Cults</a>.</p>
@@ -130,7 +154,7 @@
                             </div>
 
                             <div class="step-level-1">
-                                <h2 class="mb-4">Préparer son impression</h2>
+                                <h2 id="step2" class="mb-4">Préparer son impression</h2>
                                 <p>Il nous faut maintenant configurer notre impression dans un logiciel de tranchage ou slicer en anglais.</p>
                                 <p>Nous utilisons le logiciel libre Cura. Il est disponible sur les ordinateurs du fablab mais vous pouvez également le télécharger et l’installer sur votre ordinateur personnel pour préparer vos impressions chez vous. Ce logiciel est compatible avec Windows, Mac OS et Linux.</p>
                                 <p>Lancez le logiciel Cura, reconnaissable à son icône bleue.</p>
@@ -141,7 +165,7 @@
                                 <img class="img-fluid mb-6" src="<?php echo base_url('assets/img/toolnotice/i3d-cura-action-open.png'); ?>">
                                 
                                 <div class="step-level-2 mt-8">
-                                    <h3 class="mb-4">Éditer votre objet</h3>
+                                    <h3 id="step2-1" class="mb-4">Éditer votre objet</h3>
                                     <p>Votre objet s'affiche au centre, dans une représentation de l'espace d’impression.</p>
                                     <p>Vous pouvez ajuster la vue pour plus de confort. Vous pouvez faire tourner la vue en glissant votre pointeur tout en appuyant sur le clic droit ; pour zoomer, utilisez la roulette de la souris.</p>
                                     <p>Vous pouvez également afficher des vues prédéfinies (de face, du haut, de la gauche, de la droite…) via les icônes en bas de la fenêtre.</p>
@@ -160,7 +184,7 @@
                                 </div>
 
                                 <div class="step-level-2 mt-8">
-                                    <h3 class="mb-4">Régler les paramètres d’impression</h3>
+                                    <h3 id="step2-2" class="mb-4">Régler les paramètres d’impression</h3>
                                     <p>Sélectionnez le bon matériau, à savoir du PLA. Pour cela, allez dans le menu <span class="notice-ui-input">Matériau > Générique > PLA</span></p>
                                     <img class="img-fluid mb-6" src="<?php echo base_url('assets/img/toolnotice/i3d-cura-select-material.png'); ?>">
                                     <p>Vous pouvez ensuite régler les paramètres d’impression dans le menu dédié. Un certain nombre de préréglages vous sont proposés. Choisissez en un selon ce que vous souhaitez faire : une impression rapide, précise… Le logiciel configure les différents paramètres en conséquence.</p>
@@ -187,7 +211,7 @@
                                 </div>
 
                                 <div class="step-level-2 mt-8">
-                                    <h3 class="mb-4">Finaliser le fichier d’impression</h3>
+                                    <h3 id="step2-3" class="mb-4">Finaliser le fichier d’impression</h3>
                                     <p>Lorsque vous avez paramétré votre impression, cliquez sur <span class="notice-ui-button">Découper</span></p>
                                     <img class="img-fluid mb-6" src="<?php echo base_url('assets/img/toolnotice/i3d-cura-action-slice.png'); ?>">
                                     <p>Le logiciel va estimer le temps d’impression ainsi que la matière nécessaire. Vous pouvez jouer à nouveau sur la taille de votre objet ou ses paramètres d’impression pour réduire un peu le temps, si vous le jugez trop long.</p>
@@ -199,12 +223,12 @@
                             </div>
 
                             <div class="step-level-1">
-                                <h2 class="mb-4">Imprimer son objet</h2>
+                                <h2 id="step3" class="mb-4">Imprimer son objet</h2>
                                 <p>Allumez la machine sur le bouton situé en façade.</p>
                                 <p>L’écran s’allume. La navigation dans les menus de réglages de l’imprimante se fait à l’aide du bouton métallique. Tourner le bouton pour naviguer entre les menus ; appuyer sur le bouton pour sélectionner un menu.</p>
 
                                 <div class="step-level-2 mt-8">
-                                    <h3 class="mb-4">Installer son filament</h3>
+                                    <h3 id="step3-1" class="mb-4">Installer son filament</h3>
                                     <p>Il vous faut maintenant placer votre filament. Sur l’écran, allez dans le menu <span class="notice-ui-input">Préparer > Préchauffage PLA</span>. La buse se met alors à chauffer pour atteindre une température de 180°.</p>
                                     <div class="notice-bloc-security mb-4">Il est formellement interdit de toucher la buse lorsque l’imprimante est allumée, au risque de se brûler.</div>
                                     <p>Si un fil est déjà chargé, appuyez sur la gâchette située derrière l’imprimante, poussez légèrement le fil puis tirez-le pour le retirer. </p>
@@ -215,7 +239,7 @@
                                 </div>
 
                                 <div class="step-level-2 mt-8">
-                                    <h3 class="mb-4">Lancer son impression</h3>
+                                    <h3 id="step3-2" class="mb-4">Lancer son impression</h3>
                                     <p>Sur l’écran, sélectionnez <span class="notice-ui-input">Imprimer depuis SD</span> puis votre fichier dont le nom devrait apparaître dans la liste. </p>
                                     <p>L’imprimante va progressivement monter en température et la tête va palper le plateau. Ceci fait, l’impression se lance. Vérifiez que les premières couches s’impriment bien. Vous pouvez ensuite laisser l’impression se dérouler.</p>
                                     <div class="notice-bloc-security mb-4">Lors de vos manipulations sur l’imprimante, il est recommandé d’éviter les objets pendants (cravates, pendentifs…) et de s’attacher les cheveux. Ceux ci pourraient être entraînés par la tête d’impression.</div>
@@ -223,7 +247,7 @@
                                 </div>
 
                                 <div class="step-level-2 mt-8">
-                                    <h3 class="mb-4">Récupération de l’impression</h3>
+                                    <h3 id="step3-3" class="mb-4">Récupération de l’impression</h3>
                                     <p>L’impression se termine normalement. Décollez la pièce avec une spatule, en insérant l’outil entre le plateau et le bord de l’objet et en donnant un petit coup sec.</p>
                                     <p>Avant d’éteindre la machine, vous devez impérativement faire refroidir la buse. Pour cela, sur l’écran de contrôle, sélectionnez <span class="notice-ui-input">Préparer > Refroidir</span></p>
                                     <div class="notice-bloc-security mb-4">N'éteignez sous aucun prétexte la machine tant que le ventilateur est en marche ou que la température de la buse est supérieur à 50°C, au risque d'endommager l'imprimante.</div>