+++ assets/scripts/progressbar.js
@@ -1,0 +1,10 @@
+$(document).ready(function() {
+    var docHeight = $(document).height(),
+    windowHeight = $(window).height(),
+    scrollPercent;
+
+    $(window).scroll(function() {
+        scrollPercent = $(window).scrollTop() / (docHeight - windowHeight) * 100;
+        $('.progressBar').width(scrollPercent + '%');
+    });
+});