+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 0.1.0 - 17/04/2020
+
+New header layout
+New layout on tool pages
+Add a ToC on notice page (desktop version only)
+
 ## Version 0.0.2 - 15/04/2020
 
 Update the disclamer on tool pages